Get waveforms to work in the video

+++ src/com/asofterspace/assMusician/video/elements/Waveform.java
@@ -25,8 +25,8 @@ public class Waveform {
 		for (int x = 0; x < width; x++) {
 			int loudMax = 0;
 			int loudMin = 0;
-			for (int i = ((step + x - (width/2)) * leftData.length) / totalFrameAmount; i < ((1 + step + x - (width/2)) * leftData.length) / totalFrameAmount; i++) {
-				int offset = i;
+			for (long i = ((step + x - (width/2)) * (long) leftData.length) / totalFrameAmount; i < ((1 + step + x - (width/2)) * (long) leftData.length) / totalFrameAmount; i++) {
+				int offset = (int) i;
 				if ((offset >= 0) && (offset < leftData.length)) {
 					if (leftData[offset] > loudMax) {
 						loudMax = leftData[offset];
@@ -39,7 +39,6 @@ public class Waveform {
 			int top = vertPos - ((loudMax * 64) / (8*16*16*16));
 			int bottom = vertPos - ((loudMin * 64) / (8*16*16*16));
 			img.drawLine(x, top, x, bottom, foregroundColor);
-			// System.out.println("x: " + x + ", top: " + top + ", bottom: " + bottom);
 		}
 	}
 }
